backend/requirements.txt
flask
flask-cors
numpy
pandas
scikit-learn
joblib
pytesseract
pillow
